requirements.txt
bs4
langchain-openai
openai
pandas
python-dotenv